Fixed real_profit data type in details of futures orders

+++ pkg/model/linearswap/getdetailinformationoforderresponse.go
@@ -44,7 +44,7 @@ type DetailInformationOfOrder struct {
 	MarginAccount   string   `json:"margin_account"`
 	MarginMode      string   `json:"margin_mode"`
 	IsTpsl          int      `json:"is_tpsl"`
-	RealProfit      int      `json:"real_profit"`
+	RealProfit      float64  `json:"real_profit"`
 	ReduceOnly      int      `json:"reduce_only"`
 	Trades          []*Trade `json:"trades"`
 }
